Added the "packages" module
Added the "packages" module to make it easier to add Puppet Classes
to the project without having to create a bunch of additional
repositories. It is meant to be temporary, though, since specific
modules using the Puppet Librarian mechanism is the "right" way to
do it. But during active development of the Class this is a lot
more convenient.